# resources/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "URL Linker",
      "description": "Dynamically link URLs with custom names for shortening or improved readability.",
      "version": "0.0.1",
      "license": {
        "name": "MIT",
        "url": "https://mit-license.org"
      }
    },
    "host": "g3t.ca",
    "basePath": "/api/v1",
    "schemes": ["https"],
    "paths": {
      "/urls": {
        "post": {
          "summary": "Creates a new URL alias",
          "description": "",
          "operationId": "createUrl",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "key",
              "in": "body",
              "description": "alias of URL",
              "required": true,
              "type": "string"
            },
            {
              "name": "url",
              "in": "body",
              "description": "URL to redirect to",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "schema": {
                "$ref": "#/definitions/Url"
              }
            }
          },
          "security": ["client_id" , "client_secret"]
        },
        "get": {
            "summary": "Returns all URL alias objects",
            "description": "",
            "operationId": "getUrls",
            "consumes": [],
            "produces": ["application/json"],
            "parameters": [],
            "responses": {
              "200": {
                "schema": {
                    "values": {
                        "$ref": "#/definitions/Url"
                    }
                }
              }
            },
            "security": ["client_id" , "client_secret"]
        }
      },
      "/urls/{key}": {
        "get": {
            "summary": "Returns the URL alias object with the matching key",
            "description": "",
            "operationId": "getUrl",
            "consumes": [],
            "produces": ["application/json"],
            "parameters": [
              {
                "name": "key",
                "in": "path",
                "description": "URL key to operate on",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "schema": {
                  "$ref": "#/definitions/Url"
                }
              }
            },
            "security": ["client_id" , "client_secret"]
        },
        "put": {
          "summary": "Updates a URL alias",
          "description": "",
          "operationId": "updateUrl",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "key",
              "in": "path",
              "description": "URL key to operate on",
              "required": true,
              "type": "string"
            },
            {
              "name": "key",
              "in": "body",
              "description": "alias of URL",
              "required": false,
              "type": "string"
            },
            {
              "name": "url",
              "in": "body",
              "description": "URL to redirect to",
              "required": false,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "schema": {
                "$ref": "#/definitions/Url"
              }
            }
          },
          "security": ["client_id" , "client_secret"]
        },
        "delete": {
            "summary": "Deletes a URL alias",
            "description": "",
            "operationId": "deleteUrl",
            "consumes": [],
            "produces": [],
            "parameters": [
              {
                "name": "key",
                "in": "path",
                "description": "URL key to operate on",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "204": {}
            },
            "security": ["client_id" , "client_secret"]
          }
      },
      "/users/self": {
        "get": {
            "summary": "Returns the current user",
            "description": "",
            "operationId": "getSelf",
            "consumes": [],
            "produces": ["application/json"],
            "responses": {
              "200": {
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            },
            "security": ["client_id" , "client_secret"]
        },
        "put": {
          "summary": "Updates the current user",
          "description": "",
          "operationId": "updateSelf",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "clientSecret",
              "in": "body",
              "description": "New client secret",
              "required": false,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "security": ["client_id" , "client_secret"]
        }
      }
    },
    "securityDefinitions": {
      "client_id": {
        "type": "apiKey",
        "in": "header",
        "name": "g3t-client-id"
      },
      "client_secret": {
        "type": "apiKey",
        "in": "header",
        "name": "g3t-client-secret"
      }
    },
    "definitions": {
      "Url": {
          "type": "object",
          "properties": {
              "key": {
                  "type": "string"
              },
              "url": {
                  "type": "string"
              },
              "userId": {
                  "type": "integer",
                  "format": "int32"
              }
          }
      },
      "User": {
          "type": "object",
          "properties": {
              "id": {
                  "type": "integer",
                  "format": "int32"
              },
              "clientId": {
                  "type": "string"
              },
              "isAdmin": {
                  "type": "boolean"
              }
          }
      }
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    }
  }